split pages into smaller components
